+++ Tower Defence Beta Version/Assets/Scripts/UI/Enemy/EnemyHealthPoint.cs
@@ -1,13 +1,13 @@
+using System;
 using DG.Tweening;
 using UnityEngine;
 public class EnemyHealthPoint : HealthBar
 {
     [SerializeField] private Enemy _enemy;
-
+    
     private void OnEnable() => _enemy.ChangedHealthPoint += OnChangeHealth;
 
     private void OnDisable() => _enemy.ChangedHealthPoint -= OnChangeHealth;
 
-    //TODO UI FOR ENEMY
-    public override void OnChangeHealth(float health) => Slider.DOValue(health, FillDuration);
+    public override void OnChangeHealth(int health) => Slider.DOValue(health, FillDuration);
 }
